// src/client/__generated__/types.ts
export interface GetOrder_getOrderByNumber_customer {
  __typename: "Customer";
  id: string;
  name: string | null;
  identificationNumber: string | null;
}

export interface GetOrder_getOrderByNumber_createdBy {
  __typename: "User";
  id: string;
  name: string | null;
}

export interface GetOrder_getOrderByNumber_items_material {
  __typename: "Material";
  id: string;
  name: string;
}

export interface GetOrder_getOrderByNumber_items {
  __typename: "OrderItem";
  id: string;
  material: GetOrder_getOrderByNumber_items_material | null;
  name: string | null;
  width: number | null;
  height: number | null;
  pieces: number | null;
  printedPieces: number;
  producedPieces: number;
  totalPrice: number;
  totalTax: number;
}

export interface GetOrder_getOrderByNumber {
  __typename: "Order";
  id: string;
  number: number;
  customer: GetOrder_getOrderByNumber_customer | null;
  status: OrderStatus;
  note: string | null;
  urgency: number;
  totalPrice: number;
  totalTax: number;
  createdAt: any;
  updatedAt: any;
  createdBy: GetOrder_getOrderByNumber_createdBy;
  items: GetOrder_getOrderByNumber_items[];
}

export interface GetOrder {
  getOrderByNumber: GetOrder_getOrderByNumber | null;
}

export interface GetOrderVariables {
  number: number;
}

/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: UpdateOrder
// ====================================================

export interface UpdateOrder_updateOrder_customer {
  __typename: "Customer";
  id: string;
  name: string | null;
  identificationNumber: string | null;
}

export interface UpdateOrder_updateOrder_createdBy {
  __typename: "User";
  id: string;
  name: string | null;
}

export interface UpdateOrder_updateOrder_items_material {
  __typename: "Material";
  id: string;
  name: string;
}

export interface UpdateOrder_updateOrder_items {
  __typename: "OrderItem";
  id: string;
  material: UpdateOrder_updateOrder_items_material | null;
  name: string | null;
  width: number | null;
  height: number | null;
  pieces: number | null;
  printedPieces: number;
  producedPieces: number;
  totalPrice: number;
  totalTax: number;
}

export interface UpdateOrder_updateOrder {
  __typename: "Order";
  id: string;
  number: number;
  customer: UpdateOrder_updateOrder_customer | null;
  status: OrderStatus;
  note: string | null;
  urgency: number;
  totalPrice: number;
  totalTax: number;
  createdAt: any;
  updatedAt: any;
  createdBy: UpdateOrder_updateOrder_createdBy;
  items: UpdateOrder_updateOrder_items[];
}

export interface UpdateOrder {
  updateOrder: UpdateOrder_updateOrder;
}

export interface UpdateOrderVariables {
  id: string;
  input: UpdateOrderInput;
}

/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: DeleteOrder
// ====================================================

export interface DeleteOrder_deleteOrder {
  __typename: "Order";
  id: string;
}

export interface DeleteOrder {
  deleteOrder: DeleteOrder_deleteOrder;
}

export interface DeleteOrderVariables {
  id: string;
}
// ...
export enum OrderStatus {
  DONE = "DONE",
  NEW = "NEW",
  READY_TO_PRINT = "READY_TO_PRINT",
  TO_BE_SHIPPED = "TO_BE_SHIPPED",
  WAITING_FOR_CALCULATION = "WAITING_FOR_CALCULATION",
  WAITING_FOR_PRODUCTION = "WAITING_FOR_PRODUCTION",
}
// ...
export interface UpdateOrderInput {
  customerId?: string | null;
  items?: UpdateOrderItemInput[] | null;
  note?: string | null;
  status?: OrderStatus | null;
  totalPrice?: number | null;
  totalTax?: number | null;
  urgency?: number | null;
}

export interface UpdateOrderItemInput {
  height?: number | null;
  id?: string | null;
  materialId?: string | null;
  name?: string | null;
  pieces?: number | null;
  totalPrice: number;
  totalTax: number;
  width?: number | null;
}